fr_core/requirements.txt
CMake
Click==7.0
dlib==19.18.0
face-recognition==1.2.3
face-recognition-models==0.3.0
numpy==1.18.1
pytesseract
Pillow